% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsts_create.R
\name{bsts_create}
\alias{bsts_create}
\title{Create a bsts_model}
\usage{
bsts_create(df, date_variable, response, nseasons, niter = 100,
  target_variable, group_variable, model.options = BstsOptions(),
  rebag_vars = FALSE, inclusion_probability = 0.01)
}
\arguments{
\item{df}{Dataframe}

\item{date_variable}{Column name containing date}

\item{response}{Group that is being forecast}

\item{nseasons}{Seasonality of data - will be assumed if not provided}

\item{niter}{Number of MCMC iterations to attempt}

\item{target_variable}{Column name containing metric to be forecast}

\item{group_variable}{Column name indicating group names - will build into regressors}

\item{model.options}{Additional options for BSTS}

\item{rebag_vars}{Create pseudo aggregate variable of other regressors}

\item{inclusion_probability}{Minimum probability of inclusion in final model to show in returned predictors}
}
\value{
list object - bsts.model, data.frame of predictors
}
\description{
Create a bsts_model
}
