% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsts_test.R
\name{bsts_test}
\alias{bsts_test}
\title{Test a BSTS_Model - Highest Function level from bstsTest package}
\usage{
bsts_test(df = df, validate_n = 20, date_variable, response, na_fill = 0,
  nseasons = 7, trend_type, data_frequency, niter = 100, target_variable,
  group_variable, model.options = BstsOptions(), rebag_vars = FALSE,
  rebag_mean_vars = FALSE, inclusion_probability = 0.1)
}
\arguments{
\item{df}{Dataframe}

\item{validate_n}{N Rows of the df will be used for cross valdiation}

\item{date_variable}{Name of date variable column - will not accept lazy}

\item{response}{Response variable name - will not accept lazy}

\item{na_fill}{- Fill for NA rows - defaults to 0}

\item{nseasons}{- Seasonality of data - defaults to 7 - also takes "auto" to detect seasons automatically}

\item{trend_type}{- If using auto for nseasons, specify what type of seasonality exists [multiplicative/additive]}

\item{data_frequency}{- If using auto for nseasons, specify the frequency of the data.}

\item{niter}{- Iterations to run BSTS}

\item{target_variable}{- Name of target variable column}

\item{group_variable}{- Name of column to group by}

\item{model.options}{- Any extra BSTS Options to add in using BstsOptions()}

\item{rebag_vars}{- T/F - Sum all variables into a new predictor.}

\item{rebag_mean_vars}{- T/F Take the mean of all variables into a new prediction.}

\item{inclusion_probability}{Floor for including variables in output - doesn't affect model right now.}
}
\description{
Test a BSTS_Model - Highest Function level from bstsTest package
}
