% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mape.R
\name{MAPE}
\alias{MAPE}
\title{MAPE}
\usage{
MAPE(actual, predicted)
}
\arguments{
\item{actual}{vector or list of reported/response values}

\item{predicted}{vector or list of predicted values}
}
\description{
Mean Absolute Percentage Error of Forecast against Response
}
