% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_multi.R
\name{select_multi}
\alias{select_multi}
\title{Rapid Function to select without worrying about data.table/frame or anything else}
\usage{
select_multi(df, cols)
}
\arguments{
\item{df}{Dataframe}

\item{cols}{Character vector of columns to select}
}
\description{
Rapid Function to select without worrying about data.table/frame or anything else
}
